call_assistant_backend/requirements.txt
Flask
google-genai
python-dotenv
# opencv-python # Removed - No longer needed for video
pyaudio
# Pillow # Removed - No longer needed for video/image processing
# mss # Removed - No longer needed for screen capture
taskgroup; python_version < "3.11"
exceptiongroup; python_version < "3.11"
